+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Add the use of alias parameter with test for them in ConvertFrom-PSRDiacritics
+
 ## [0.3.0-preview0001] - 2024-02-26
 
 ### Fixed